make getTemplate and updateTemplate methods

See merge request khudaikoff/simple-project-initilizer!43

+++ backend/src/main/java/ccfit/nsu/ru/spi/repository/TemplateRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
-public interface TemplateRepository extends JpaRepository<TemplateEntity, Double> {
-
+public interface TemplateRepository extends JpaRepository<TemplateEntity, Long> {
     List<TemplateEntity> findAll();
 
 }
